requirements.txt
apache-airflow
psycopg2-binary
requests
pandas